add navigation bar and implement routing
