+++ src/App.js
@@ -1,5 +1,12 @@
 import React, { Component } from 'react';
+import {
+  Route, BrowserRouter, Routes,
+} from 'react-router-dom';
 import Calculator from './components/Calculator';
+import Home from './components/Home';
+import Navbar from './components/Navbar';
+import './components/Navbar.css';
+import Quote from './components/Quote';
 
 export default class App extends Component {
   constructor(props) {
@@ -9,9 +16,14 @@ export default class App extends Component {
 
   render() {
     return (
-      <div className="App">
-        <Calculator />
-      </div>
+      <BrowserRouter>
+        <Navbar />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/calculator" element={<Calculator />} />
+          <Route path="/quote" element={<Quote />} />
+        </Routes>
+      </BrowserRouter>
     );
   }
 }
