+++ src/components/Home.css
@@ -1,0 +1,8 @@
+.home {
+  margin-top: 150px;
+  padding-left: 50px;
+}
+
+.home p {
+  font-size: 18px;
+}
