+++ src/components/Home.jsx
@@ -1,0 +1,90 @@
+import React from 'react';
+import './Home.css';
+
+const Home = () => (
+  <div className="home">
+    <h1>Welcome to our page!</h1>
+    <p>
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      <br />
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      <br />
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      <br />
+      <br />
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      <br />
+      <br />
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+      Lorem ipsum dolor sit amet consectetur adipisicing elit.
+      Ducimus suscipit, omnis eius pariatur libero ipsum.
+      Magnam sint quisquam, corrupti eius vero ipsa ullam! Optio consectetur magnam nihil vero,
+      quidem laboriosam!
+    </p>
+  </div>
+);
+
+export default Home;
