+++ src/components/Quote.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import './Quote.css';
+
+const Quote = () => (
+  <div className="quote">
+    <p>Mathematics consists of proving the most obvious thing in the least obvious way.</p>
+  </div>
+);
+
+export default Quote;
